split tests by class; add grid.boundary to __init__, README, replication test
